Implement Claude API client
I still have to write tests for this client, as well as the API request
and response structs.

There is a design question to resolve here, too. In cogito-openai, the
individual submodels are pub, whereas here, they are not pub; and in
cogito-openai, there are doctests for some of the simpler submodels, but
we cannot have doc tests for non-pub structs.

We should be consistent across crates, so either we should make these
submodels pub and write doctests for them, or we should make the
submodels in cogito-openai non-pub and then necessarily change the tests
from doctests to regular tests (which might be better for performance
reasons anyway).

I hate to pollute the API with unnecessary structs, but some of the
submodels, such as ClaudeUsage, might actually be useful to consumers of
this crate.

But then I also don't feel like going through and writing a ton of
simple accessor methods, so if these structs do become pub, then maybe
the attributes should just be pub as well; but I kind of hate that
because then what if I want to write more complicated accessors
eventually? Pub attributes might be the "Rust way", but I think they
violate encapsulation, which is a policy necessary for good code no
matter what the language. Also, if attributes for submodels are pub
here, they should be pub in cogito-openai, too, for consistency...but I
already wrote all those nice accessor methods in cogito-openai, and I
hate to get rid of them.

I will punt that decision until I start writing tests for this code.