Wrap ClaudeResponse::content in accessor method
A private accessor method makes it a little easier to work with, as it
can return an iter() instead of forcing code to explicitly call
ClaudeResponse::content.iter() every time it needs to use the content.

I made this private to avoid confusion with ClaudeResponse::result(),
which is what callers really should be using. I should make
OpenAIResponse::output() and OpenAIResponse::concatenate() private to
match, but I have to update a number of integration tests if I do that,
and I will have to think about what those are really testing before I
modify them.