OpenAI submodels should be private
The submodels aren't really necessary except to mirror the output from
OpenAI -- OpenAIResponse does all the heavy lifting -- and we shouldn't
pollute the public API with a bunch of spurious data structures
representing implementation details.

Doing so meant that I had to move some doctests to regular ol' tests,
because the relevant data structure is now private and so can't be used
in doctests, but that's okay because regular ol' unit tests compile and
run faster anyway.